Enable account selection for MMI codes that involve "dialing".
Some MMI codes (specifically a small set of GSM codes) involve pressing
"dial" to send the code. The connection service identifies that this is
an MMI code and responds accordingly, including canceling the call. In
the multi-SIM case, we want to bring up a dialog to select a SIM
"connect" through. We can use the current PRE_DIAL_WAIT state and select
account dialog to do this.

Bug: 17573147
Change-Id: Id765e595f3149c088ae5de72a97cc2034338b6ee