Creating connections for conference event package participants
- Add "addExistingCall" method to Connection service wrapper to handle
the creation of new calls for connections which already exist in telephony.
- Fixed but in setIsConference message where an isValidCallId check made
outside of the handler would result in a newly added call NOT being
marked as conferenced due to a race condition.  The MSG_SET_IS_CONFERENCED
handler already checks if the call exists before performing its operation.

Bug: 18057361
Change-Id: I2348c8e6985cd831f2cae56b37dfcb33bae014a8

Make add-call a global property of telecom. (2/4)

ADD_CALL didn't make sense as a property of Connection or Call.
This changes it to be a global property instead.

Bug: 18285352
Change-Id: I5189e114e74dba2d81d74c6d24cf5b17f1e0922a

Set appropriate capabilities to TestConference

This allows the "Manage conference" screen to show up correctly
for test conference calls

Bug: 18269622
Change-Id: I8467654a8158c5da8c354057f613f122b22a0ed3

Create IMSConference to represent IMS conference calls.

- Removing check in ConnectionServiceWrapper which would prevent adding
a new confernece with no calls in it; an IMS conference by default will
have no calls.

Bug: 18200934
Change-Id: I2e70b9350e779310e8c4ee8e91bc8020b126bcfa

Change-Id: I2e70b9350e779310e8c4ee8e91bc8020b126bcfa