(Telecom-system part 2) Move telecom intialization from application to service.
Instead of having TelecomApp initialize the global state of telecom, it
will bind to the telecom service and the service itself will
initialize the global state. This prepares telecom to run in the system
process since at that point TelecomAppw will stop running and the
framework will need to use the telecom service to start the global state.

Bug: 18112269

Conflicts:
	src/com/android/server/telecom/TelecomServiceImpl.java

Change-Id: Ib29fc579633bcfb86a4b7953565dd66f3546e77e