Remove the LCH on sub having active call.
When the DIAL request on a sub fails due to some reason
and if there is a sub which is in lch having a active call
it will continue to be in lch without this change reason being
that incall ui will not be launched for such cases.
So in order to handle those cases check for any sub in lch and release
it.

Change-Id: I5de92db4e0fefd2814de390cf1707f793d767761
CRs-Fixed: 778958,761725,764489