IMS Incorrect Call State Check For Answering Call
Participant is getting dropped from conference call when an incoming
call is answered on top of the held conference call. This is due
to a child call's being checked as part of the check to ensure
that a ringing call is answered only if an active and held call
doesn't already exist.

CRs-Fixed: 816127
Change-Id: I0001f9b2be971bb7e1f020971f4917a1488077e3