Telecom: Don't treat PRE_DIAL_WAIT call as live call
The PRE_DIAL_WAIT call is treated as live call, after adding a new ring
call to the call set in CallsManager, in updateForgroundCall method,
because the call set is unordered, sometimes the new forground call is
updated to ring call, in this case, the ring sound can be set normally,
sometimes, the forground call is updated to the PRE_DIAL_WAIT call,
in this case, ring sound can not be set properly.

Don't treat PRE_DIAL_WAIT call as live call.

Change-Id: I8f218ddf8ab80de380bdeddf117a7d5d347ec18c
CRs-Fixed: 911121