IMS-VT: Refresh Add call option when Video call downgrades to VoLTE
"android.telecom.extra.DISABLE_ADD_CALL" extra guards the visibility
of "Add Call" option. The option will not be shown if the extra is
present and vice-versa. When call is downgraded, the extra is removed
but is not refreshing the "Add Call" option. Fix this by refreshing
"Add Call" option when the extra is removed.

Change-Id: I95cf794bd4a7af6c395b1fb35295008346fdb36a
CRs-Fixed: 2151650